backend/requirements.txt
setuptools==73.0.1
django==5.1
djangorestframework==3.15.2
psycopg==3.2.1
drf-spectacular==0.27.2
djangorestframework-simplejwt==5.3.1
